Move the RoomEditModal to an edit sub-directory
Also apply minor refactorings in the same component.
