+++ indico/modules/rb/client/js/common/rooms/edit/RoomPhoto.module.scss
