+++ indico/modules/rb/client/js/common/rooms/index.js
@@ -10,7 +10,7 @@ import * as selectors from './selectors';
 
 export {default as reducer} from './reducers';
 export {default as RoomDetailsPreloader} from './RoomDetailsPreloader';
-export {default as RoomEditModal} from './RoomEditModal';
+export {default as RoomEditModal} from './edit/RoomEditModal';
 export {default as RoomRenderer} from './RoomRenderer';
 export {default as modalHandlers} from './modals';
 export {actions, selectors};
